added dev mode for cli

+++ src/dashboard/_cli/_src/bin/psi-catalog.ts
@@ -7,7 +7,6 @@ import {CatalogModel} from '../models/Catalog'
 import {ArchiveModel} from '../models/Archive'
 import UriManager from '../lib/UriManager'
 import MongoConnection from '../lib/MongoConnection'
-
 // program
 //     .command('default', {isDefault: true})
 //     .description('Default command, for whatever reason')
@@ -20,14 +19,16 @@ program
     .description('Path to the storm data')
     .option('-p, --path <type>','Give path to the catalogs',undefined)
     .option('-a, --all','Add all directories of --path as a catalog',false)
+    .option('-d, --dev','dev mode',false)
     .action(async (cmd) => {
         const {
             path,
             all,
+            dev
         } = cmd
 
         //make sure a path is given
-        if(!path) return console.log('Please provide a path')
+        if(!path) return colorize.warning('Please provide a path')
 
         //connect to db
         const uriManager = new UriManager();
@@ -35,7 +36,9 @@ program
         await mongoConnection.connect()
 
         //TEST COMMAND
-            await CatalogModel.deleteMany({})
+            if(dev) {
+                await CatalogModel.deleteMany({})
+            }
             //await ArchiveModel.deleteMany({})
             //await ImageModel.deleteMany({})
 
@@ -49,20 +52,6 @@ program
             colorize.info(catalogResult.message)
         }
         await mongoConnection.close()
-
-        //if archive option is true, then create the archives
-        // if(addArchives) {
-        //     colorize.log(`Inserting archives for the ${catalogResult.data.length} created catalogs`)
-        //     for(let i =0;i<catalogResult.data.length;i++) {
-        //         const element = catalogResult.data[i]
-        //         const {
-        //             name,path,_id
-        //         } = element
- 
-        //         await archive.addArchives(path,_id,{all:allArchives})
-
-        //     }
-        // }
     })
